server-api: Add MutationApiController
