+++ packages/public/server/src/module/Api/src/Factory/ApiControllerFactory.php
@@ -29,8 +29,6 @@ use Api\ApiManager;
 use Api\Controller\ApiController;
 use Api\Service\AuthorizationService;
 use Common\Factory\AbstractControllerFactory;
-use Discussion\DiscussionManager;
-use Discussion\DiscussionManagerInterface;
 use Instance\Manager\InstanceManager;
 use Instance\Manager\InstanceManagerInterface;
 use License\Manager\LicenseManager;
@@ -60,10 +58,6 @@ class ApiControllerFactory extends AbstractControllerFactory
         /** @var ApiManager $apiManager */
         $apiManager = $serviceManager->get(ApiManager::class);
         $controller->setApiManager($apiManager);
-
-        /** @var DiscussionManagerInterface $discussionManager */
-        $discussionManager = $serviceManager->get(DiscussionManager::class);
-        $controller->setDiscussionManager($discussionManager);
 
         /** @var LicenseManagerInterface $licenseManager */
         $licenseManager = $serviceManager->get(LicenseManager::class);
